invoices

+++ src/Database/migrations/000081_create_order_items_table.php
@@ -8,6 +8,9 @@ return new class extends Migration
 {
     public function up(): void
     {
+        if (Schema::hasTable('order_items')) {
+            return;
+        }
         Schema::create('order_items', function (Blueprint $table) {
             $table->id();
             $table->foreignId('order_id')->constrained()->onDelete('cascade');
@@ -16,11 +19,14 @@ return new class extends Migration
             $table->string('sku');
             $table->string('name');
             $table->integer('quantity');
+            $table->integer('returned_quantity')->default(0);
             $table->decimal('price', 10, 2);
             $table->decimal('final_price', 10, 2);
             $table->decimal('promo_code_discount_amount', 10, 2)->default(0);
             $table->decimal('total', 10, 2);
             $table->decimal('tax_amount', 10, 2)->default(0);
+            $table->string('tax_rate')->nullable();
+            $table->string('unit')->nullable();
             $table->timestamps();
 
             $table->index(['order_id', 'product_id']);
